[photon-core] [NFC] Code Cleanup, spelling, and grammar (#945)

+++ photon-core/src/main/java/org/photonvision/vision/aruco/ArucoDetectorParams.java
@@ -48,9 +48,9 @@ public class ArucoDetectorParams {
 
         logger.info("Setting decimation from " + m_decimate + " to " + decimate);
 
-        // We only need to mutate the parameters -- the detector keeps a poitner to the parameters
+        // We only need to mutate the parameters -- the detector keeps a pointer to the parameters
         // object internally, so it should automatically update
-        parameters.set_aprilTagQuadDecimate((float) decimate);
+        parameters.set_aprilTagQuadDecimate(decimate);
         m_decimate = decimate;
     }
